+++ packages/core/src/Models/Product.php
@@ -216,8 +216,9 @@ class Product extends BaseModel implements SpatieHasMedia
     public function getSearchableAttributes()
     {
         $attributes = $this->getAttributes();
+        $keys = array_merge(['attribute_data'], config('lunar.search.exclude_model_attributes.product', []));
 
-        $data = Arr::except($attributes, 'attribute_data');
+        $data = Arr::except($attributes, $keys);
 
         foreach ($this->attribute_data ?? [] as $field => $value) {
             if ($value instanceof TranslatedText) {